Split the Constants(added in 65bdc13) into a separate module. (enthought#301)
